Major design milestone.
Framework types will be for now created on demand, because we now
implement equality and identity on the Wrapper base class by using the
handle.  And since they keep no state, we always can recreate safely.

The subclassed types on the other hand do keep state, and those we
want to track.  So when we surface objects from a pointer, we find a
live object if needed

Also fixes the lifecycle bits for classes.

See comments for details.