Finally happy with collections.
Clean up the support for collections and other bits.

1. Built-in types now surface a zero property, that can be used to
initialize those instances, this will be needed to not hard-code the
initializer for future support of the larger-data sized versions of
Godot.

2. Make the GodotVariant protocol require an init constructor from a
variant, it forces the user to add a required constructor, but I tried
too many alterantives, and none worked: I need a way for the
GodotCollection to instantiate the subclases of Object.

3. Found a user for the liveFramerowkrTypes: to surface
the-most-derived type in situations like adding some nodes to a
collection, and then retrieving them.

4. Replaced a lot of manual code with generated code.