Change of strategy for object lookups.
The previous attempt with GodotCollection started to surface some of
the requirements, but porting a game made clear where the API
downsides are.

APIs like "Node.getNode (...) -> Node" currently were surfacing
objects of type Node when they were first surfaced to the Swift world,
because this is what the type signature is, but these are typically
subclasses.

What I want is an API that would return the actual type (that happens
to be a subclass of Node).   This patch is a preparation to achieve
that goal.

The end-goal is to figure out given a handle to a native objects what
the most derived type is.  So we will eventually need a `lookupObject`
that not only checks the surfaced objects, but can construct a new
wrapper on-demand, and this would require a new required conformance
into a `init (nativeHandle:)` on subclasses of Wrapped.   This would
become very ugly, having two mandatory and obscure requirements.

So I decided to split GodotCollection into two: VariantCollection for
the Godot "built-in" types, and ObjectCollection for anything that
subclasses the Godot Object.   That way, I could eliminate one of the
conformance requirements on user subclasses.

The next step is to have lookupObject actually determine the most
derived type, and then call the appropriate constructor.