Update on: Change of strategy for object lookups.
This implements the support to have lookupObject actually determine
the most derived type, and then call the appropriate constructor.

This will allow APIs like "Node.getNode (...) -> Node" currently were
surfacing objects of type Node when they were first surfaced to the
Swift world, because this is what the type signature is, but these are
typically subclasses.

There is quite a lot of room for optimization here, along the lines of
not using GStrings and use StringNames, and likely not using this huge
type descriptor table