Add load and print family of utilities in GD
The load function works similarly to that of GD.Load in the C# variant
and is a means of being consistent with the GDScript preload function.

Three new print methods are also available, which function similarly to
Swift's traditional print function. These methods also handle converting
the items accordingly to ensure they get represented properly, as well
as converting the resulting string to the Variant type that the normal
GD.print methods require.