API usability.
There are plenty of methods in the generated binding that produce things like:

```
func addSomething (something: Int, with: Int)
```

And users would need to call this with:

```
   object.addSomething (something: 3, with: 4)
```

I now changed this to drop the first argument name in those conditions
to avoid the repetition.  I chose this over dropping the suffix of the
method as it felt like it was more clear.

So now the above would produce:

```
func addSomething (_ something: Int, with: Int)
```

Now users can use:

```
   object.addSomethign (3, with: 4)
```