More works towards removing the incorrect usage of pointers and moving towards using withUnsafePointer everywhere.
Additionally, because it is so difficult to find the problems in the
sea of warnings, I started to reduce some warnings related to the use
of let vs var in the results variables.

The annoying part is that we need to use "var" for things like "Int",
"Float", and various struct value, but we need to use "let" in cases
where we have a class and we pass a pointer to the content (like
"result.content").

There are also mild warning fixes, like extension methods do not need
to be declared public when they are part of a public extension.