When invoking virtual methods and surfacing instances of objects, rather than passing the instance of a class, if the class has subclasses, now we find the most-derived type.
This means that rather than creating an `InputEvent` that we pass to
`Node._input (event:)`, we now create the most precise one, depending
on the actual type of the object (like `InputEventMouseButton`, or
`InputEventMouseMotion`).

This process is only done for Godot objects that have subclasses, for
those that do not, we just create the straight wrapper (we already had
support for user-defined subclasses).

Fixes