Fixes the problem where we would register every Godot-virtual method as being implemented in Swift, even if it was not.
This was initially reported as: "Can't interact with custom controls"
here: https://github.com/migueldeicaza/SwiftGodot/issues/198

The problem was that we were providing a default implementation for
"_has_point" in every class, which had a default implementation that
returned false.

Godot does not provide a base-class method that can be invoked
(different languages deal with this in different ways, I had wired it
up to return a default value, other languages abort, we perhaps should
do that).

Instead, Godot checks if one of its virtual methods has been
registered, and only calls if that is the case, but if not, it
provides the default behavior.

This change is based on the work of mgshorter@gmail.com and Padraig O
Cinneide on the changes done to the baseclasses and the macro itself.