Make @SceneTree path argument optional (#297)
* Add .nova to gitignore

I use Nova to edit Swift packages nowadays, but I dont want my personal
settings getting pushed up. Hence, I am adding it to the gitignore.

* Make SceneTree path argument optional

If the path argument is not provided in the macro, it will use the
variable declarations identifier text instead of throwing an error. For
example:

var char2D: CharacterBody2D?

would look for a node named char2D exactly. This approach makes no
assumptions on spacing or case; it is up to the user to ensure their
name matches when using the default argument approach. Otherwise, the
path argument can be used to explicitly set formatting or pathing.