Try to fix the Variant leak problem and add surface a way of calling unref.
Variant leak: this brings back the Variant memory leak fix, with one crash
fix to go along with it - but given that at least Kev is still reporting a
crash, it is clear there is at least another scenario not covered.   So
I am turning this on for now, and if someone runs into a crash, they can
disable it while we extract the test case.

This also includes Patrick Beard's new method to surface the `unref` operation
on RefCounted, that we need to get the semantics of unref right.   This will
eventually go into RefCounted