Start of Strict Concurrency support for SwiftGodot.
The goal is to make the library useful to developers that might use
multiple threads/tasks with Godot, and have the Swift compiler assist
them, and help them make errors.  So it is not as important to get
this done quickly, as it is to be useful.

This first set of changes are fairly straight forward, they enable
strict concurrency, which causes some 30,000+ warnings to be flagged.
Most of these errors were related to shared global state, so this
change addresses a big chunk of those 30,000+ instances by using `let`
in a few places rather than `var`.

The second set of changes are related to a few globals in SwiftGodot
and how we handle them.  A batch of nonisoalted values are the lock
around the liveFrameworkObjects and liveSubtypedObjects.  These are
nonisolated, precisely, becuase they are protected manually by that
lock.

The EntryPoint's library, token and gi are global variables that are
used to operate Godot, and currently are only set when initializing
Godot.  I do not intent to put additional checks to prevent the user
from not initializing Godot.

That said, there is a bug #72 for future improvements for those
globals, and in that case, we might change this

Other changes:

* Our generated OptionSets are flagged as Sendable, as they are.

* Removed some variables that we had exposed and did not have to be.