Export enumeration support
The idea is to do this in two steps, the first step is to use the
convenience @PickerNameProvider macro on an enumeration, like this:

```
enum MyEnumeration: Int, CaseIterable {
    case first
    case second
}
```

And then to export this, use the `.enum` parameter to Export:

```
@Godot
class Demo: Node {
     @Export(.enum)
     var myState: MyEnumeration
}
```

One limitation of the current change is that this works by using
`CaseIterable`, either by manually typing it, or using the
`PickerNameProvider` macro, this means that the built-in enumerations
from SwiftGodot currently are not surfaced, I could look into that.

Additionally, this does not support the string enumeration values that
Godot supports, that could be easily added by probing if the values
are strings, and using a different mechanism to format the key/values
passed to the property registration.