Adds support for generating the structures in "native_structures" in the JSon file, and also surfaces a handful of methods that take pointers.
Some work is in place now, currently it maps "const void*" and
"AudioFrame*" to OpaquePointer, but this is not ideal.

What I need is to handle these in a case-by-case basis, so I can
expose better types when surfacing this.

For example, the AudioFrame APIs while they pass a pointer, in the
cases where they are used there is a separate parameter that indicates
the number of frames in place. So this really should be a
UnsafeMutableBufferPointer<AudioFrame> that has been initialized with
the base address and the number of elements, but this would require
custom code for each case I suspect.

And you can see that sometimes the values are return values, so they
should not be buffers:

https://gist.github.com/migueldeicaza/28974ecb3a1bdfffa818a0f0b6999cb0