fix second leak of #513
This patch adds a test for the second leak described in #513, and fixes
the leak.

Without this patch, the leak happens because `Variant.asObject` has an
unneeded call to `rc.reference()` which increments the `RefCounted`
object's reference count. As far as I can tell, nothing balances this
increment.

`Variant.asObject` calls `lookupObject(nativeHandle:)`, which returns
a Swift wrapper for an object whose reference count has already been
incremented (if the object is `RefCounted`). The Swift wrapper
balances that increment with a decrement in its `deinit`.