Add tests cases for, and fix, #513 (#524)
* add some support for using a local libgodot build

* fix better the first leak of #513

The changes in d1d6ca4 made several `Geometry2DTests` fail.

The actual unbalanced retain causing the leak in
`VariantRepresentable.swift`:

``` swift
extension ContentVariantRepresentable {
    public init? (_ variant: Variant) {
        guard Self.godotType == variant.gtype else { return nil }

        var content = Self.zero
        withUnsafeMutablePointer(to: &content) { ptr in
            variant.toType(Self.godotType, dest: ptr) // <-- THIS IS THE UNBALANCED RETAIN
        }

        self.init(content: content)
    }
}
```

The `variant.toType` function calls into Godot to copy out the built-in
type stored in `variant`. If that built-in type has an internal
reference count, Godot increments it. Then, the call to
`self.init(content: content)` increments the count again. The retain
performed by `variant.toType` is never balanced.

This patch fixes the bug by generating a new
`init(alreadyOwnedContent:)` initializer for each Godot builtin class
type. This needs to be on the builtin wrappers (like `GDictionary`,
`GArray`, etc.), rather than on `Variant` which is where d1d6ca4 put it.

This patch also adds a test case to check for the leak by looking at
Godot's `MEMORY_STATIC` performance counter, which is only enabled if
Godot was built with `DEBUG_ENABLED`.

* fix second leak of #513

This patch adds a test for the second leak described in #513, and fixes
the leak.

Without this patch, the leak happens because `Variant.asObject` has an
unneeded call to `rc.reference()` which increments the `RefCounted`
object's reference count. As far as I can tell, nothing balances this
increment.

`Variant.asObject` calls `lookupObject(nativeHandle:)`, which returns
a Swift wrapper for an object whose reference count has already been
incremented (if the object is `RefCounted`). The Swift wrapper
balances that increment with a decrement in its `deinit`.

---------

Co-authored-by: Rob Mayoff <mayoff@dqd.com>