Allocation-less marshaling for built-in and class types methods, and `Callable` changes (#533)
* Add Package.resolved to .gitignore

SPM does it this way so I gues it's fine?

* Add .vscode folder to .gitignore

We don't have any substantial configurations for .vscode specifically and yet it's a common IDE to work from on Swift on Windows to justify it

* Add benchmarks and save their baseline for Apple M3 Max

* Create Arguments struct

The idea behind this struct is that if a `call` is initiated from Swift side, it just wraps a user-created array of [Variant]s
If a `call` is  issued from Godot, the marshaled arguments storage is already managed by Godot, so we can just lazily retrieve the variants inside without an extra allocation of array

* Introduce UnsafePointer helper functions generator and tests with their usages and assumptions checks

* Bring a test for `Variant.call`

* Reimplement `Variant.call`

Since arguments payload is transient and doesn't outlive the scope of the function, using `withUnsafeTemporaryAllocation` is a much better solution that ideally fits the case. A tandem of an array and an eager `allocate` are an overkill for this scenario.

* Replace `[Variant]`-based API with `borrowing Arguments`.

This change should be source-compatible for most of the cases and has an easy workaround in case people indeed want their own persistent copy of `borrowing Arguments`.

* Remove a space between @c_decl and args.

The latest (Swift 6) toolchain prints a warning for those

* Add an apparent way to opt-of out new marshaling implementation

* Modify the comment to add some extra insight

* Adapt ClassServices function registration to use optimized path of `Arguments`

If Godot makes a call into Swift land, there is no need to allocate the arguments storage array proactively. Just use Godot-managed one and retrieve Variants lazily inside Arguments implementation.

For FunctionInfo, use struct instead of a class. Class unmanaged runtime is more sophisticated due to refcounting/managing control block/etc, Godot doesn't need to do any ref-counting on it, so it's just not worth it.

* Treat `CallableWrapper` similarly to `FunctionInfo`

* Update generator implementation to use new unsafe pointer nesting strategy

Temporary arrays are ditched, all pointers are kept on stack, using `UnsafeRawPointersN*`. Update baseline for benchmarks.

* Ditch extra stack copy

Since we already make a copy during args preparation, there is no need to nest unsafe pointers in this approach.

* Simplify the code for current approach. Do minor renaming. Move `withMemoryRebound` to the helper function itself.

* Fix the documentation

* Try Windows

* Revert "Try Windows"

This reverts commit babfa08d72c70a97904c72e0bb18890257a9e2d3.

* Update .gitignore

* Make two separate tokens in the builder

* Remove redundant `isVararg` check

* Fallback to old marshalling on non macOS

* Include all Apple platforms

* Add @inline(__always) to withUnsafeArgumentsPointer

* Fix occasional build error due to C stdlib functions not resolved

* Make generator runtime opt-in for legacy marshaling

* Add removed old API, mark it as deprecated

* Inline storage

* Add var arg methods performance test and baseline for it

* Use const pointer in gi.variant_new_copy.

src is GDExtensionConstVariantPtr, aka const void *, ditch redundant init

* Fix vararg setup indentation

* Separate MethodGen implementations to Legacy and New

* Fix indentation