Parameterize the "minimum_initialization_level" for the extension
While "scene" level is going to be the right choice for the majority of projects, .editor might be a more appropriate choice for an Extension which is purely an in-editor tool. Likewise, an Extension that has to add Core or Server types might need the earlier settings. So, make this a parameter on initializeSwiftModule() with a default value of .scene.