Preparation for 4.4 API update.
From Gabor Koncz: Previously assert(method.isVirtual) (along with its
assert (!method.isVirtual) pair) expressed the precondition that a
method is virtual if and only if it has a method hash. However, as
previously discussed here, there were changes in the API generation
which resulted in all methods having a method hash, which made the
precondition no longer apply. This meant that to separate the two
cases, we no longer check the existence of a method hash, but we
directly check the isVirtual boolean, and so the assertions there
would be redundant.

There have been some changes to the api generation in 4.4; one of
these is that all methods (including virtual ones) have a hashcode
now, so we cannot use that information to determine virtual-ness in
the SwiftGodot proxy generation. These changes were probably made in
godot because of the addition of is_required, which is also why we now
call the godot method in the default implementation of the generated
virtual methods on the Swift side. This is also why we need these
changes here, to handle the case when a result is assigned in the case
of a virtual call (the result type will be an OpaquePointer there, so
we don't have to specify .handle in that

This patch does not yet include the new JSON file for the API or the
new .h file