One challenge is that including a simple type can end up bringing a lot of external types, and this is problematic for the split.  In particular, it is problematic for SwiftGodotRuntime which will only contain a handful of types that are useful for interop.
I took the liberty of attempting to bring Node, and that unleashed a
world of dependencies, mostly in return types and property types.

Node would surface Viewport and Window and some others.  I tried to
bring those, but each one kept bringing more stuff, and was defeating
the purpose of SwiftGodotRuntime.

So instead, I did this: the low-hanging fruit were return values, like
Viewport or Window, and for those, the generator can now produce a
modified version of the API that returns the raw handle and is only
exposed via @_spi(SwiftGodotPrivate), so the methods are in the class,
but just not exposed.  So for a method getFoo() -> Foo, we would now
generate 'rawGetFoo() -> OpaquePointer' in SwiftGodotRuntime, and then
provide the actual 'getFoo()' as an extension in the other library.

Then when the actual types are present (in SwiftGodotCore, but we
might end up extending this to others), then an extension method will
produce the missing strong binding by using the underlying "raw" method.