Barebones-split.
This is based on the code from the `split` branch, but rebased on the
warning-fixed main branch, and with a smaller scope than `split`.

To recap: `split` attempts to create multiple dynamic libraries for SwiftGodot,
but the issue that we are running into is that SwiftPM currently forces
replicating copies of any dependencies defined on the same project one each
split library.   So if you have SwiftGodot2D that depends on SwiftGodotCore, an
entire copy of SwiftGodotCore is embedded into SwiftGodot2D, and if you have
SwiftGodot3D it will also embed its own copy of SwiftGodotCore, if you use both
2D and 3D you end up with two independent copies of SwiftGodotCore.

Here, I took a slightly different approach, I only create two libraries:
SwiftGodotRuntime, intended by those that just want the basic interop features,
and SwiftGodot that references the above and brings everything else.   Given the
current semantics of SwiftPM, this is not different from the public API
perspective than the old version.