Updated all the tests to the new names and new code generation, everything works, except the one bit where I had to remove the method inference for the Node.self, as I do not have a way of testing for that now.
I suspect I can add an extension in SwiftGodot for Node that would handle this,
but I am tripping into an Xcode issue that is generating continuously all my
code for "indexing" into my project directory every couple of seconds, and it
brings Xcode down to a halt - will explore later.   Even typing one line can
take one to two minutes as I wait for the various beachballs and Xcode to
regenerate and reparse all of this data.