Fixes looking up the most-derived type.
After we split SwiftGodot in two, the built-in "godotFrameworkCtors" array that
used to be source of all known Godot types, we broke in subtle ways, as we would
not produce most-derived types in a variety of situations (Nathan reported this
problem and so did Chris).

But really, it could hit anyone at any time.

Rather than trying to sort out a new registration system, this was a good time
to drop the list of arrays and use the dynamic lookup system that I had tried
earlier on (#555 - but when I last tried that, I attempted two things at once,
and the second part was flawed, so we bring back the good part).